[Bronze II] Title: Hashing, Time: 128 ms, Memory: 14164 KB -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -7,10 +7,10 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int L = Integer.parseInt(br.readLine());
         String s = br.readLine();
-        int result = 0;
+        long result = 0;
 
         for (int i = 0; i < L; ++i){
-            result += (int)(s.charAt(i) - 'a' + 1) * (Math.pow(31, i));
+            result += (long)(s.charAt(i) - 'a' + 1) * (Math.pow(31, i));
         }
         System.out.println(result);
     }
